Move SystemTestClients to fixtures package
