+++ java/system-test/system-test-suite/src/main/java/sleeper/systemtest/suite/dsl/SystemTestReporting.java
@@ -24,6 +24,7 @@ import sleeper.systemtest.drivers.partitioning.PartitionReportDriver;
 import sleeper.systemtest.drivers.util.TestContext;
 import sleeper.systemtest.suite.dsl.reports.SystemTestCompactionJobsReport;
 import sleeper.systemtest.suite.dsl.reports.SystemTestIngestJobsReport;
+import sleeper.systemtest.suite.fixtures.SystemTestClients;
 
 public class SystemTestReporting {
