+++ java/system-test/system-test-suite/src/main/java/sleeper/systemtest/suite/dsl/ingest/SystemTestIngest.java
@@ -23,7 +23,7 @@ import sleeper.systemtest.drivers.ingest.IngestByQueueDriver;
 import sleeper.systemtest.drivers.ingest.IngestSourceFilesDriver;
 import sleeper.systemtest.drivers.instance.SleeperInstanceContext;
 import sleeper.systemtest.drivers.util.WaitForJobsDriver;
-import sleeper.systemtest.suite.dsl.SystemTestClients;
+import sleeper.systemtest.suite.fixtures.SystemTestClients;
 
 import java.nio.file.Path;
